api_data/requirements.txt
fastapi==0.111.0
jose==1.0.0
pandas==2.2.2
pydantic==2.7.4
pytest==8.1.1
python-dotenv==1.0.1
python_jose==3.3.0
SQLAlchemy==2.0.29
uvicorn==0.30.1
pyodbc
passlib